Added REcipe Index page

+++ app/controllers/users_controller.rb
@@ -1,6 +1,6 @@
 class UsersController < ApplicationController
-	before_filter :signed_in_user, only: [:create]
-	before_filter :admin_user, only: :destroy
+	#before_filter :signed_in_user, only: [:create]
+	#before_filter :admin_user, only: :destroy
 	
   def index
   	@users = User.all
@@ -20,10 +20,10 @@ class UsersController < ApplicationController
   	if params[:create]
   		@user = User.new(params[:user])
   		if @user.save
+  			flash[:success] = "New User created."
   			redirect_to index_path  			
-  		else
-  			render 'new'
-  			flash[:notice] = "A problem occurred. PANIC!"
+  		else  					
+  			render 'new'  			
   		end
   	elsif params[:cancel]
   		redirect_to index_path
@@ -40,11 +40,10 @@ class UsersController < ApplicationController
 		if params[:save]
   		@user = User.find(params[:id])
   		if @user.update_attributes(params[:user])
-  			flash[:success] = "Profile updated"
+  			flash[:success] = "Profile updated."
   			redirect_to index_path  			
   		else
-  			render 'edit'
-  			flash[:notice] = "A problem occurred. PANIC!"
+  			render 'edit'  			
   		end
   	elsif params[:cancel]
   		redirect_to index_path
@@ -52,22 +51,20 @@ class UsersController < ApplicationController
   end
 
 
-	def destroy		
-		@users = User.find(params[:user_ids])
-		@users.each do |user|
-		 	user.destroy
+	def destroy
+		if params[:user_ids].nil?
+			flash[:error] = "No user/s selected."
+			redirect_to index_path
+		else
+			@users = User.find(params[:user_ids])
+			@users.each do |user|
+				user.destroy
+			end
+			#User.find(params[:user_id]).destroy
+			# @user.destroy
+			flash[:notice] = "User(s) was/were successfully deleted."
+			redirect_to index_path
 		end
-
-
-		#User.find(params[:user_id]).destroy
-		# @user.destroy
-			
-		flash[:notice] = "User was successfully deleted."
-		redirect_to index_path
-		#else
-			#flash[:notice] = "User was successfully."
-			#redirect_to index_path
-		#end
 	end
 
 	#def destroy_multiple
